feat: put summary in a dedicated tab. clicking files opens explorer
